requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.0
easyocr==1.7.1
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
fonttools==4.43.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
httptools==0.6.0
idna==3.4
imageio==2.31.5
imutils==0.5.4
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.26.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
psutil==5.9.5
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
thop==0.1.1.post2209072238
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.198
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
setuptools>=65.5.1
